Update blackboxsok2024.md
Some minor text edits. Also, should we include the table that shows the under-explored settings in our taxonomy? The text there alone is a bit more boring. Either we cut the text that broadly describes the under-explored settings and only keep the top-k one, or include both the whole table and also the corresponding texts.